Time: 4 ms (84.58%) | Memory: 8.1 MB (78.91%) - LeetSync

+++ 9-palindrome-number/palindrome-number.cpp
@@ -7,7 +7,7 @@ long temp = x;
     if (x < 0) return false;
     while(temp != 0)   
     {  
-    long remainder = temp % 10;  
+    int remainder = temp % 10;  
     reverse = reverse * 10 + remainder;  
     temp = temp/10;  
     }
